use mtime as a 64 bit value

+++ src/database.h
@@ -34,7 +34,7 @@ void vh_database_file_data_delete (database_t *database, const char *file);
 void vh_database_file_grab_insert (database_t *database, file_data_t *data);
 void vh_database_file_grab_update (database_t *database, file_data_t *data);
 void vh_database_file_grab_delete (database_t *database, const char *file);
-int vh_database_file_get_mtime (database_t *db, const char *file);
+int64_t vh_database_file_get_mtime (database_t *db, const char *file);
 void vh_database_file_get_grabber (database_t *database,
                                    const char *file, list_t **l);
 void vh_database_file_insert_dlcontext (database_t *database,
